Aalborg Air sensor code
Working setup for the Aalborg project. The sketch interacts with the sensor and saves data on a logger, while sending it on to TTN. The libraries: "Arduino-lmic-master" and "opcn2 - Arduino-master" have been modified for this sketch. You WILL need to use the modified libraries for the sketch to work properly.